# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMPADU Keuangan API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/keuangan/pembayaran": {
            "get": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Ambil semua data pembayaran",
                "operationId": "9e254db6eb04b2dc046de45e5924f1da",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Simpan data pembayaran",
                "operationId": "609e63d5d804e1167b4fa406c18df0b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nim",
                                    "kategori_ukt_id",
                                    "tanggal_pembayaran",
                                    "jumlah",
                                    "metode_pembayaran"
                                ],
                                "properties": {
                                    "nim": {
                                        "type": "string"
                                    },
                                    "kategori_ukt_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_pembayaran": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "metode_pembayaran": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil disimpan"
                    }
                }
            }
        },
        "/api/keuangan/pembayaran/{id}": {
            "get": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Lihat detail pembayaran",
                "operationId": "6c6e067999e83c2ffdb6a7d4b2b4feb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail ditemukan"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Update pembayaran",
                "operationId": "497e404d65f2dd2d1181a1d3e75e4042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nim": {
                                        "type": "string"
                                    },
                                    "kategori_ukt_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_pembayaran": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "jumlah": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "metode_pembayaran": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Hapus pembayaran",
                "operationId": "e7d77f8822ddc2b78111518ebb99284d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/api/keu-tagihan": {
            "get": {
                "tags": [
                    "Keu Tagihan"
                ],
                "summary": "Menampilkan semua data tagihan",
                "operationId": "890c28958053869c5a441479248e43b6",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Keu Tagihan"
                ],
                "summary": "Membuat tagihan baru",
                "operationId": "de04348be5e08883bc8494791616fbf7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_mahasiswa",
                                    "id_jenis_tagihan",
                                    "jumlah",
                                    "id_tahun_ajar"
                                ],
                                "properties": {
                                    "id_mahasiswa": {
                                        "type": "integer"
                                    },
                                    "id_jenis_tagihan": {
                                        "type": "integer"
                                    },
                                    "jumlah": {
                                        "type": "number"
                                    },
                                    "id_tahun_ajar": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil disimpan"
                    }
                }
            }
        },
        "/api/keu-tagihan/{id}": {
            "get": {
                "tags": [
                    "Keu Tagihan"
                ],
                "summary": "Menampilkan detail tagihan",
                "operationId": "4c331ac876eec5830b094f652c092788",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil"
                    }
                }
            },
            "put": {
                "tags": [
                    "Keu Tagihan"
                ],
                "summary": "Memperbarui data tagihan",
                "operationId": "f5713b8f040e61609bfcea09a225cfb8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "jumlah": {
                                        "type": "number"
                                    },
                                    "id_jenis_tagihan": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Keu Tagihan"
                ],
                "summary": "Menghapus tagihan",
                "operationId": "36beedea31143be252633ab2aad2797b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Berhasil dihapus"
                    }
                }
            }
        },
...
    },
...
}